src/test/java/ch10/Ex4Test.java
package ch10;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Ex4Test {
    static Stream<Arguments> testCases() {
        return Stream.of(
                Arguments.of(0, 0),
                Arguments.of(1, 1),
                Arguments.of(2, 1),
                Arguments.of(3, 2),
                Arguments.of(4, 3),
                Arguments.of(5, 5),
                Arguments.of(6, 8),
                Arguments.of(7, 13),
                Arguments.of(8, 21),
                Arguments.of(9, 34),
                Arguments.of(10, 55)
        );
    }

    @ParameterizedTest
    @MethodSource("testCases")
    void returnExpectedResults(int n, int expectedResult) {
        int result = Ex4.parallelFibonacci(n);

        assertEquals(expectedResult, result);
    }
}